[eclipse/xtext#1629] moved xtend not to prefer xtend
