+++ org.eclipse.xtend.ide.common/src/org/eclipse/xtend/ide/common/XtendIdeModule.java
@@ -1,11 +1,11 @@
 /*******************************************************************************
- * Copyright (c) 2010, 2017 itemis AG (http://www.itemis.eu) and others.
+ * Copyright (c) 2010, 2020 itemis AG (http://www.itemis.eu) and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  *******************************************************************************/
-package org.eclipse.xtend.ide.common
+package org.eclipse.xtend.ide.common;
 
 
 /**
@@ -13,5 +13,5 @@ package org.eclipse.xtend.ide.common
  * 
  * @since 2.12
  */
-class XtendIdeModule extends AbstractXtendIdeModule {
+public class XtendIdeModule extends AbstractXtendIdeModule {
 }
