--- org.eclipse.xtend.ide.common/xtend-gen/org/eclipse/xtend/ide/common/XtendIdeModule.java
@@ -1,19 +1,0 @@
-/**
- * Copyright (c) 2010, 2017 itemis AG (http://www.itemis.eu) and others.
- * All rights reserved. This program and the accompanying materials
- * are made available under the terms of the Eclipse Public License v1.0
- * which accompanies this distribution, and is available at
- * http://www.eclipse.org/legal/epl-v10.html
- */
-package org.eclipse.xtend.ide.common;
-
-import org.eclipse.xtend.ide.common.AbstractXtendIdeModule;
-
-/**
- * Use this class to register ide components.
- * 
- * @since 2.12
- */
-@SuppressWarnings("all")
-public class XtendIdeModule extends AbstractXtendIdeModule {
-}
